Pygmy psuedo-buffs
So, pygmies suck. One of the reasons for that is that they only have access to 7 gold units, when there literally exists a 5 gold pygmy base. That one is excluded from the the pygmy only filter, and others where pygmies take a prominent role and only poses_pygmy_elite is used. This will drasticlly increase the viability of pygmies, perhaps to the level where they are not only for the insane and/or masochistic.

So basically there are regular pygmies, and 1 pygmy archer and 1 pygmy elte archer instead of everyone being elite.