Undead Debuffs
So Dustwalkers were already 47 rec points, upping that to 57 doesn't do much. Meanwhile Nadir are still being cavernwights. So that calls for a major change, an increase of 30 rec point, as well as a 37/55 increase in gold. Now your size 3, chill 3, siege strength, strength 17, amphibious, insane MR, 0 encumbrance undead are potentially able to be unbanned in all servers.  This nerf may not be enough.