Sacred pose generation weighting fix & further balance tweaks
Found another significant bug in sacred generation: sacred poses are supposed to be weighted by role and ranged sacreds are supposed to be used only as a last resort for a nation's first sacred, but a logic error was resulting in these weights being stripped out and all available poses being picked interchangeably. Fixed logic to account for role weighting, also reinserted a very low chance for ranged sacreds to be selected on purpose for a nation's first sacred.

-Further tweaks to weighting on new sacred filters
-Fixed lingering vaetti loincloth graphic transparency issue from rimvaettir addition