Switched overcharged10 and 20 to 2 and 4
The ability works differently as the original devs seemed to intent, judging by the messages in https://github.com/elmokki/nationgen/issues/804 (overcharged was presented there as #overcharged [stun dmg], rather than direct damage).

Overcharged 1 is capped at 1 damage, as Corpse Constructs do in vanilla. Higher values of it do not get capped, and so do DRN + overcharge vs. DRN AN damage. Overcharged 2 and 4 are reasonable values that are effective but can also be simply countered by shock res 5, which is in reach of most blesses or earth and air mages.