Fixed elemental secondtmpshapes so they don't shrink into random units
The new shrinkhp behaviour in Dominions 6 means that the new monsters that are left behind by some sacreds would cycle to the following modded monsters, rather than to the intended elemental shapes. This is because the monster would be given its own id, not the id of the elemental that it's copystating from.

To fix it, we copy everything from the elemental form, use clearspec to remove its special abilities like shrinkhp, then readd them without it.