Clockwork template updates to dom6
The way the clockwork armours were written was to grant them built-in weapons and on top of that have a random number of hand slots, given through hand items.

This worked in Dominions 5 because units were not required to have actual hand slots defined to be able to wield weapons. However in Dominions 6 this results in them not using the weapons that they were not given enough slots to use, even non-commander units.

On top of that, some of the heavy clockwork armours could end up dual wielding 2-handed weapons because the hand slots were irrelevant, but in Dominions 6 they'd end up using only one of the two weapons if they only have two hands.

Now hand slots are guaranteed to always be 2, and the 2-handed weapons given to heavy armours have been customized to be 1-handed instead, but with the same stats, effectively still being able to dual-wield them, but without needing to have 4 hand slots.