Reworked #additionalgods GodGen logic and added all realmless gods
Thoroughly commented GodGen, and changed the arbitrary 0 to 4 addition of individual gods to a nation to a guaranteed 1 + reduced chances for every subsequent one, so that a nation can potentially roll a large amount of them if they get very lucky (always based on their race/pantheons).

Templates for individual gods include all the ones without a realm. Some can get duplicated right now, i.e. the Dracolich, as there is a version of it that belongs to a realm and one that does not which is very minorly different.

These templates can be refined to include more detailed chances based on nation themes and races.