Fixed incorrect reference to tags when getting Unit's #itemslot
Foulspawn templates which use dynamic arm items that grant variable numbers of #itemslot for their hands were not applied correctly, as the tags array was always empty.

This was not a problem in Dominions 5 because having 0 hands did not get in the way of using their given weapons. In Dominions 6 this results in the units having useless kicks and weapons they can't use.