Explicitly gave all mount form definitions a #gcost and #rcost
The reason for this is twofold:

1) When using copystats to define the stats of a mount form, if the monster selected to copystats from is a summon, it will cost 0 gold and 0 resources, whereas if it's also a recruitable (i.e. Abysian Salamanders which cost 50g a piece), it will also copy that cost. This obscured behaviour is undesirable and hard to track down, so it's better practice to explicitly define what the mount form should cost.

2) The gcost and rcost given to the mount form is what Dominions will assign as a cost to the mount itself. When going in-game and opening a unit's stat page, you can press Ctrl+i to open its pricing details. On the mounted unit tab, it will give the gold/res cost of the unit without counting the mount form. On the mount form tab, it will give the price of the mount form alone. This distinction probably has in-game ramifications, perhaps tied to the mounted unit recruitment after riders without a mount have come home, so it's better to respect it, rather than assign all the pricing to the mounted unit as a whole.

The other reason why I find this system to be better is that it allows us to declare a standard cost for a mount form, and then to go into the item that attaches the mount form to the rider (with the #mountmnr command) and give a variation of the cost, like #gcost -5 or #gcost +10, so that some specific kinds of nations might have that mount form as cheaper or more expensive.

An example of the above is that I made most of the pygmy templates have a discounted cost for most of the mount forms that they use, to give them a very small buff and make it more thematic.

The code changes are formatting, as well as the removal of the "specifiedcost" check in MountUnit.java, as it's unused for mounts and prevented the #gcost commands on mount forms to be written into the mod files.