Sacred slider affects power calculation at the top level
Other fixes to it were localizing the power change to downstream functions that made it so it was only taken into account for some specific calculations but not others.

Ideally we should split the sacred power slider into a power slider and a cost slider, since a single sacred power slider has to be arbitrary as to whether those more powerful sacreds cost more and are less recruit everywhere (which makes sacreds more powerful but costlier), or whether it should add power for free.