Fixed Ettin helmets to work around the hard limit of three armors
As per the Dom6 manual, units have a limit of three #armor tags, which made ettins with two helmets, an armor and a shield lose their equipped shield.

However, since Dom6 does not handle multiple heads as independent hit locations, just equipping a single helmet with #armor will grant its protection to both heads and still allow to keep the shield. The gameid of the otherhelmet has been set to -1 so that they still generate a sprite, but do not get added as #armor.