Fixed Foulspawn cavalry not pricing in their mounts
This forced a paradigm change - the reason why they were not being priced is because the .dm file is created with the mounted unit with its own gcost, and the mount monster with its own gcost, and Dominions was left to combine those into the final recruitment gold cost.

Since foulspawn are template units that don't get resolved until they are recruited, and the placeholder cavalry unit that you recruit does not actually have a mount, it was never adding any mount gcost, just the averaged cost of all the unit templates as seen in Unit.java lines 1034-1050.

This changes when the mount item data gets loaded, and it gets added to the Unit instance for convenience, so that when the foulspawn generation is averaging cost of the different unit shapes it will also check what that unit's mount is, and include that price.