Fix to spells getting duplicated
The mod has to redefine spells multiple times because of Dom6's limitations with the #restricted keyword (in dom6, only 12 of these tags can be given to a custom spell). The workaround for this in NatGen is to simply define the spells again and again with different sets of restricted nations.

The issue with the above is that the mod is using #copyspell and referencing the spell by name, and then giving the spell a #name with the same words; essentially overwriting that spell. And on the next new definition of the spell with a new block of #restricted tags, it's copying the spell by name again - meaning it's copying the spell it already defined before.

Stupidly, as pointed out in https://github.com/elmokki/nationgen/issues/1221, the easiest fix is to add a space to the end of the spell name on each new spell definition, as this means that the #copyspell command referencing the vanilla spell by name will be rightfully copying the vanilla spell, and not the rewritten NatGen one.

A better solution would be to change the list of custom_spells to #copyspell by id, but it's a lot of grunt work.