Added support for = operator within special commands like #chanceinc
For example, #chanceinc "command #demon =1" would overwrite the earlier unit's #basechance to be equal to 1 if the unit is a demon.

Currently, a #chanceinc with a modifier of 1, without any operator (like + or -) is treated by default like an addition, so a #chanceinc "command #demon 1" means "add 1 to the #basechance if this unit is also a demon", rather than setting its #basechance to 1 as the notation might suggest.