Fixed polish() pass resolving custom item ids resulting in unresolved ones
When resolving the custom equipment of certain units during Unit.polish(), some of the items are marked with dependencies, such as #needstype rarm onehand and needstype offhand polearm.

These dependencies are resolved once the id of said item is resolved, but they go on to potentially replace previously equipped items (in fact, they'll replace them even if the item already equipped fulfills that dependency for that slot!).

The problem with this is that if those newly equipped items fell in slots that have already been passed over in the custom id equipment loop, their custom id will not be resolved after getting equipped, and the program will throw an exception when writing the unit's .dm lines.

Said exception has also been made explicit and has had a clear debugging message added, so that whenever this happens it's easy to figure out which unit of which nation with which seed and which item caused the specific problem.

As a bonus, the Item id resolution is now encapsulated into a static method of the Item class, which should make it easier for future debugging as well.