Tweaked unit rating calculation for cap only chances further
Last version was incorrectly dividing by number of filters which meant that a single power 4 filter resulted in lower rec-everywhere chances than a power 4 and power 2 filter, for example.

A sum of square powers is now used, and then it's scaled back down with a root. This means that, for example, two filters of level 2 will result in slightly lower rating than a single filter of level 4.

There is still a cumulative rating that gets added on top for every extra filter that the unit has, too.

With the current set up, a unit with a single power 2 filter will result in a rating of 1. If the unit has a 0.5 survivability rating, that means that it'll have slightly more than a 50% chance of being rec everywhere.