Temporarily replaced numbing weapon effect for ao-oni's cold
There is no direct equivalent in Dominions 6 to the effect id 686 in Dominions 5, which used to be an aoe1 armor negating 10 cold fatigue damage. This would have to be made into a custom weapon but then that would require adding id resolution code to the part of the item that gets the #secondaryeffect attached to it.