Fixed how filters get picked for sacreds to even probabilities across powers
The filters were being picked by power; i.e. if the unit could afford filters up to power 4, it would generate a number from 1 to 4 at random, and then it would pick a filter from those with said power.

This meant that something like dragonblood that is #power 4 would be way too common, because it would have a 25% chance to select 4 as the power for the filter, and then there are barely any #power 4 filters in the pool to contend with it.