Fixed seed generation problem with StatUpgradeManager
Because the StatUpgradeManager was using a HashMap structure to iterate through the different stats (a HashMap does not guarantee ordering!), every execution of the program using the same seed would result in slightly different results.

This isn't because of what was going on within the HashMap iteration itself, but because the HashMap iteration was adding elements to an ordered List<Filter> possibleUpgrades, meaning in every program execution the order in which it would add them would change, and thus the later EntityChances.baseChances(possibleUpgrades).getRandom(random); call would have different results.

FUTURE NOTE: to preserve seed integrity, either use data structures which guarantee ordering, or sort those structures AFTER iterating with an immutable rule.

Some such safeguards have been added here - the HashMap is now a LinkedHashMap, which guarantees ordering, and the possibleUpgrades List is sorted by stat name after the loop, as a further guarantee.