Added #chanceinc to Cynocephalian elite troop poses to allow heroes
There is a 5% chance in the HeroGenerator.java module (line 264) that a hero for a nation picks from an entirely random race. After this, it goes through the usual process of selecting a valid pose for the hero from that chosen race.

This means that the race needs to be able to handle picking poses even if they don't always have some expected themes. As they stand, all of the Cynocephalian poses are #basechance 0 by default, and have a #chanceinc to pick one or another based on whether they are tribal or civilized.

However, if a nation with other races that don't roll those themes happened to trigger that 5% chance and generate a Cynocephalian hero, then this will throw an exception, as no valid pose can be found.

The fix for now is to add an equal weight of #chanceinc moduleid herogen 1 to a few of the elite troop poses, so that if they are being chosen as a hero, they have a default chance of 1.