Reworked #weapon and #armor line writing into mod file
One of the main reasons is to be able to sort weapons by range and define #weapon with the higher ranges first. This is because of a Dominions bug where if a unit has multiple ranged weapons, it will only use the range of the topmost weapon to figure out how close to get to the enemy. This is a problem for nagas, who have a built-in, short range poison spit but can also be archers.

Because weapons are currently split into pure modded command weapons like the naga natural weapons defined in their bases directly as #weapon X, and actual Item weapons that get equipped in slots, there's some convoluted data structure handling that needs to happen here to pool those together first, then figure out their ranges, then write them into the .dm file.

The problem here is customitems with a #range command that are not vanilla are currently not being sorted like this. The new getWeaponRange() method should be expanded to fetch its #range command defined originally, through nationGen.customItemsHandler().customItems. In cases of negative #range values that scale up with strength, we can possibly return them directly, since the same strength scaling on the same unit will result in the same range. We would still have to make sure of whether lower negatives are lower or higher strength scaling