Fixed mounts not getting weapons or armor
This was a side-effect of the change in how #weapon and #armor lines get written into the modfile that was done in commit 42d6812d4d25b8fd767bdc81d302a7a20b42671f.

Because it switched MountUnits and ShapeChangeUnits to write their #armor and #weapon lines by calling writeWeaponLines() and writeArmorLines(), which are functions inherited from the Unit class, and these functions were not using this.getCommands() (which gathers all the unit's commands, be it in its own command list or the related instances like mounts and other shapes).

They were using this.commands, which will only ever contain commands directly added into Unit.commands.