Further fix to specialcommanderfilters generating with tier 3 mages
Refer to commit 21a4a79087f6d350f0210706b163ff9064e2daac for the previous fix.

Commanders that use a [#themeinc thisitemtag "tier 3" *0] type of themeinc, which affects item lists and zeroes out some of them can end up with empty possible item lists.

For example, merrow casters always get "hair" as a helmet, which is tagged as a tier 1, 2 and 3 item. If a tier 3 merrow mage generates with a special filter that has the [#themeinc thisitemtag "tier 3" *0] themeinc, this single head item will have its chance set to 0.

Currently this throws an exception. Even though this could be handled more quietly, like throwing a warning, the filter chances or item lists should arguably be tweaked so that there's no such possibility that no option is available.

The fix here is to replace the themeinc that zeroes out any tier 3 items by a tag which disables the filter for tier 3 mages entirely.